Update app Gradle config and resources
Clean up and modernize app build configuration and resources: remove AboutLibraries plugin aliases, add aboutlibraries.json (debug & main) and locales_config.xml. Add packaging.jniLibs.keepDebugSymbols for several native libs to preserve debug symbols, remove legacy multiDex/renderScript settings, and switch dev flavor resourceConfigurations to androidResources.localeFilters. Normalize project dependency notation (project(":module:...")) and apply related buildSrc/Gradle adjustments across modules.

+++ app/src/main/java/eu/kanade/tachiyomi/ui/reader/settings/NovelAppearanceView.kt
@@ -132,7 +132,7 @@ class NovelAppearanceView @JvmOverloads constructor(context: Context, attrs: Att
         swatch.background = GradientDrawable().apply {
             shape = GradientDrawable.OVAL
             setColor(color)
-            setStroke(1.dpToPx, 0x66888888.toInt())
+            setStroke(1.dpToPx, 0x66888888)
         }
     }
